Put data structures in SpecificProof.hs
