+++ lib/theory/src/OpenTheory.hs
@@ -6,7 +6,7 @@
 
 module OpenTheory (
     module OpenTheory
-    
+    ,module SpecificProof
     ) where
 
 import GHC.Generics
@@ -36,38 +36,12 @@ import           Theory.Tools.IntruderRules
 import           Term.Positions
 
 import           Utils.Misc
-import Theories
 import Safe
 
-------------------------------------------------------------------------------
--- Specific proof types
-------------------------------------------------------------------------------
+import Theories
 
--- | Proof skeletons are used to represent proofs in open theories.
-type ProofSkeleton    = Proof ()
 
--- | Convert a proof skeleton to an incremental proof without any sequent
--- annotations.
-skeletonToIncrementalProof :: ProofSkeleton -> IncrementalProof
-skeletonToIncrementalProof = fmap (fmap (const Nothing))
-
--- | Convert an incremental proof to a proof skeleton by dropping all
--- annotations.
-incrementalToSkeletonProof :: IncrementalProof -> ProofSkeleton
-incrementalToSkeletonProof = fmap (fmap (const ()))
-
--- | Proof skeletons are used to represent proofs in open theories.
-type DiffProofSkeleton    = DiffProof ()
-
--- | Convert a proof skeleton to an incremental proof without any sequent
--- annotations.
-skeletonToIncrementalDiffProof :: DiffProofSkeleton -> IncrementalDiffProof
-skeletonToIncrementalDiffProof = fmap (fmap (const Nothing))
-
--- | Convert an incremental proof to a proof skeleton by dropping all
--- annotations.
-incrementalToSkeletonDiffProof :: IncrementalDiffProof -> DiffProofSkeleton
-incrementalToSkeletonDiffProof = fmap (fmap (const ()))
+import SpecificProof
 
 
 -- Lemma construction/modification
