+++ lib/theory/src/SpecificProof.hs
@@ -1,0 +1,38 @@
+module SpecificProof (
+    module SpecificProof
+) where
+
+import           Prelude                             hiding (id, (.))
+
+import           Theory.Proof
+
+------------------------------------------------------------------------------
+-- Specific proof types
+------------------------------------------------------------------------------
+
+-- | Proof skeletons are used to represent proofs in open theories.
+type ProofSkeleton    = Proof ()
+
+-- | Convert a proof skeleton to an incremental proof without any sequent
+-- annotations.
+skeletonToIncrementalProof :: ProofSkeleton -> IncrementalProof
+skeletonToIncrementalProof = fmap (fmap (const Nothing))
+
+-- | Convert an incremental proof to a proof skeleton by dropping all
+-- annotations.
+incrementalToSkeletonProof :: IncrementalProof -> ProofSkeleton
+incrementalToSkeletonProof = fmap (fmap (const ()))
+
+-- | Proof skeletons are used to represent proofs in open theories.
+type DiffProofSkeleton    = DiffProof ()
+
+-- | Convert a proof skeleton to an incremental proof without any sequent
+-- annotations.
+skeletonToIncrementalDiffProof :: DiffProofSkeleton -> IncrementalDiffProof
+skeletonToIncrementalDiffProof = fmap (fmap (const Nothing))
+
+-- | Convert an incremental proof to a proof skeleton by dropping all
+-- annotations.
+incrementalToSkeletonDiffProof :: IncrementalDiffProof -> DiffProofSkeleton
+incrementalToSkeletonDiffProof = fmap (fmap (const ()))
+
