+++ lib/theory/tamarin-prover-theory.cabal
@@ -113,6 +113,7 @@ library
       Options
       Lemmas
       OpenTheory
+      SpecificProof
 
     other-modules:
       Theory.Text.Parser.Exceptions
